+++ Examples/DVKCompanionShowcase/DVKCompanionShowcase/App.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+import DuplexVoiceKitUI
+
+@main
+struct DVKCompanionShowcaseApp: App {
+    var body: some Scene {
+        WindowGroup { DVKCompanionStartupView() }
+    }
+}